All Taxa Calcs

Given a data frame (of all treatments and all taxa), along with a data frame specifying treatment comparisons to make.

Usage

all_taxa_calcs(
  X.all,
  comparisons,
  output_directory,
  M.soil = Sdat,
  taxon_column = "taxon",
  density_column = "Density",
  copies_ul_column = "t.copies.ul",
  tube_column = "unique.tube",
  treatment_column = "unique.tmt",
  soil_g_column = "g.dry.soil.tube",
  growth.model = "linear",
  prop.O.from.water = 0.6,
  v.frac = 30,
  copies.cell = 6,
  pgC.cell = 0.1,
  CI = 0.9,
  draws = 1000,
  tailed.test = 1
)

Arguments

X.all

data frame with data for all taxa and treatments

comparisons

data frame specifying which treatments to compare; must contain these columns: "comparisonID", "iso.compare", "trt.code.1", "trt.code.2", "trt.refs", "days"

output_directory

Directory to write the output files to

M.soil

data frame with mass of soil for each replicate; must contain two columns, one for replicate ID and one for mass of soil; default is NULL meaning that is M.soil is not provided, only WAD and ape results are returned

taxon_column

column with taxon IDs (formerly vars[1])

density_column

column with density data (formerly #vars[2])

copies_ul_column

column with abundance or copy number data (formerly vars[3])

tube_column

column with tube or sample IDs (formerly vars[4])

treatment_column

column with treatment IDs (formerly vars[5])

soil_g_column

column with soil weight (formerly vars[6])

growth.model

model to use for calculating growth ("linear" or "exponential")

prop.O.from.water

proportion of oxygen atoms in DNA that come from environmental water

v.frac

volume of a fraction within a tube (uL)

copies.cell

assumed number of 16S copies per cell

pgC.cell

assumed mass of carbon per cell (units: picograms)

CI

confidence interval (0-1)

draws

number of bootstrap iterations

tailed.test

specifies whether a 2-tailed or 1-tailed permutation test should be performed in calculating the p-value; default is 2; note: if 1-tailed is specified, then the function assumes group 2 (trt.code.2) is the 'heavy' treatment and group 1 (trt.code.1) is the 'light' treatment

Value

A list with

taxonID: identifier for the taxon (factor)

comparisonID: identifier for the treatment comparison

trt.code.1: ID code for treatment 1 (factor)

trt.code.2: ID code for treatment 1 (factor)

wad1.obs.mean: mean of observed weighted average density values for each tube (rep) in treatment 1

wad2.obs.mean: mean of observed weighted average density values for each tube (rep) in treatment 2

wad1.boot.mean: mean of bootstrapped weighted average densities treatment 1

wad1.boot.median: median of bootstrapped weighted average densities treatment 1

wad1.boot.CI.L: lower end of confidence interval of bootstrapped weighted average densities for treatment 1

wad1.boot.CI.U: upper end of confidence interval of bootstrapped weighted average densities for treatment 1

wad2.boot.mean: mean of bootstrapped weighted average densities treatment 2

wad2.boot.median: median of bootstrapped weighted average densities treatment 2

wad2.boot.CI.L: lower end of confidence interval of bootstrapped weighted average densities for treatment 2

wad2.boot.CI.U: upper end of confidence interval of bootstrapped weighted average densities for treatment 2

wad.diff.obs: observed difference in WADs between treatments (i.e. subtracting mean WAD -- across reps -- for group 1 from that for group 2)

wad.diff.boot.mean: mean of the bootstrapped differences in WAD between group 2 and group 1

wad.diff.boot.median: median of the bootstrapped differences in WAD between group 2 and group 1

wad.diff.boot.CI.L: lower end of confidence interval of the bootstrapped differences in WAD between group 2 and group 1

wad.diff.boot.CI.U: upper end of confidence interval of the bootstrapped differences in WAD between group 2 and group 1

wad.diff.p.value: P-value (obtained by permutation test); the probability of observing a difference between treatments that is as large or larger than the observed difference, if the null hypothesis (that the treatments have equal WADs) is true

ape.obs: observed excess atom fraction based on the two treatments in the comparison

ape.boot.mean: mean of the bootstrapped excess atom fraction values

ape.boot.median: median of the bootstrapped excess atom fraction values

ape.boot.CI.L: lower confidence intervals of the bootstrapped excess atom fraction values

ape.boot.CI.U: upper confidence intervals of the bootstrapped excess atom fraction values

r.obs: observed intrinsic rate of increase based on the two treatments in the comparison

r.boot.mean: mean of the bootstrapped intrinsic rate of increase values

r.boot.median: median of the bootstrapped intrinsic rate of increase values

r.boot.CI.L: lower confidence intervals of the bootstrapped intrinsic rate of increase values

r.boot.CI.U: upper confidence intervals of the bootstrapped intrinsic rate of increase values

f.obs: observed carbon flux based on the two treatments in the comparison

f.boot.mean: mean of the bootstrapped carbon flux values

f.boot.median: median of the bootstrapped carbon flux values

f.boot.CI.L: lower confidence intervals of the bootstrapped carbon flux values

f.boot.CI.U: upper confidence intervals of the bootstrapped carbon flux values

N1.obs.mean: mean of observed abundances for each tube (rep) in treatment 1

N2.obs.mean: mean of observed abundances for each tube (rep) in treatment 2

N1.boot.mean: mean of bootstrapped abundances treatment 1

N1.boot.median: median of bootstrapped abundances treatment 1

N1.boot.CI.L: lower end of confidence interval of bootstrapped abundances for treatment 1

N1.boot.CI.U: upper end of confidence interval of bootstrapped abundances for treatment 1

N2.boot.mean: mean of bootstrapped abundances treatment 2

N2.boot.median: median of bootstrapped abundances treatment 2

N2.boot.CI.L: lower end of confidence interval of bootstrapped abundances for treatment 2

N2.boot.CI.U: upper end of confidence interval of bootstrapped abundances for treatment 2

N.obs.mean: mean of observed abundances for each tube (rep) in treatments 1 and 2

N.boot.mean: mean of bootstrapped abundances treatments 1 and 2

N.boot.median: median of bootstrapped abundances treatments 1 and 2

N.boot.CI.L: lower end of confidence interval of bootstrapped abundances for treatments 1 and 2

N.boot.CI.U: upper end of confidence interval of bootstrapped abundances for treatments 1 and 2

message: warning message listing replicates in either of the 2 treatments (does not include the reference group) that had no occurrences (i.e., all y values for a rep are 0); value for message is "none" when all replicates are present (single character value)

The following files will be written

bootstrapped_wad1.txt

bootstrapped_wad2.txt

bootstrapped_wad_diff.txt

bootstrapped_ape.txt

bootstrapped_r.txt

bootstrapped_C_fluxes.txt

bootstrapped_N1.txt

bootstrapped_N2.txt

bootstrapped_N.txt

Details

For all taxa, calculate growth rates, carbon fluxes, abundances, excess atom fraction, and return statistics to assess whether the means of weighted average density (WAD) differ significantly between treatments

For the written files, values are the quantities specified in the name of the data frame, columns are individual bootstrap samples (i.e., mean of resampled data for a single iteration) and rows correspond to different taxa and treatment comparisons

Written by Ben Koch & Natasja van Gestel

Note

requires the functions 'WAD.func', 'boot.TUBE.func', 'diff.wad.calc', 'MW.calc', 'ape.calc', 'r.calc', 'f.calc', & 'comparison.message'

not implemented yet: include an argument to specify which quantities the user wants calculated e.g., ('wad.diff', 'ape', 'f', 'r', 'all')

(not all inputs would be needed for certain quantities)

If M.soil is not provided, then only WAD and ape results are returned; if M.soil is provided then all results (WAD, ape, r, f, N) are returned

treatment code names cannot contain spaces

the treatments in the comparisons data frame (columns trt.code.1, trt.code.2, trt.refs) can be separated by semicolons, commas, or spaces (or a combination of semicolons and spaces or commas and spaces)

if the comparisons data frame is imported as a csv file, using commas to separate reference treatments may be problematic unless quoting is used (recommend using tab-delimited text files for import)

this function (along with ape, r, f) cannot handle comparing doubly-labeled treatments (yet)

'copies' is really 'copies.ul'

units of abundance (N) are in copies/unit medium, where the unit medium is specified by vars[6] (i.e., mass of soil: 'g.soil' by default)